secunik/backend/requirements.txt
# backend/requirements_base.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# File Processing
python-magic==0.4.27
python-magic-bin==0.4.14  # Windows
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
py7zr==0.20.6
rarfile==4.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Networking & Async
aiofiles==23.2.1
httpx==0.25.1
requests==2.31.0

# AI Integration
openai==1.3.5
tiktoken==0.5.1
tenacity==8.2.3

# Analysis Tools (without pypff)
# pyshark==0.6  # Also might have issues
# python-registry==1.3.1  # Also might need compilation
networkx==3.2
tldextract==5.1.0

# Other
python-dateutil==2.8.2
Jinja2==3.1.2
email-validator==2.1.0
python-dotenv==1.0.0
jsonschema==4.20.0